% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisisem2025-package.R
\docType{package}
\name{analisisem2025-package}
\alias{analisisem2025}
\alias{analisisem2025-package}
\title{analisisem2025: utilidades para analisis operativo de la Encuesta Multiproposito}
\description{
Anotaciones a nivel de paquete para mantener sincronizados los imports
compartidos cuando se regenere \code{NAMESPACE} con roxygen2.
}
\author{
\strong{Maintainer}: David Gómez Lizarazú \email{degomezl@dane.gov.co}

}
\keyword{internal}
