% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{arreglar_utf8_df}
\alias{arreglar_utf8_df}
\title{Limpiar data frame completo a UTF-8}
\usage{
arreglar_utf8_df(df)
}
\arguments{
\item{df}{Data frame.}
}
\value{
Data frame limpio en UTF-8.
}
\description{
Aplica \code{arreglar_utf8()} a todas las columnas de texto de un data frame.
}
