% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizacion_cruces.R
\name{clasificar_documento_cruce}
\alias{clasificar_documento_cruce}
\title{Clasificar calidad basica de un documento para cruce}
\usage{
clasificar_documento_cruce(x)
}
\arguments{
\item{x}{Vector con documentos originales o normalizados.}
}
\value{
Vector character con la categoria de calidad documental.
}
\description{
Aplica una clasificacion operativa sobre documentos normalizados para
distinguir documentos vacios, genericos, muy cortos, compuestos solo por
ceros y potencialmente validos.
}
\examples{
clasificar_documento_cruce(c(NA, "0000", "123", "123456"))

}
