% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticos_genericos.R
\name{diagnosticar_variable_basica}
\alias{diagnosticar_variable_basica}
\title{Diagnosticar una variable basica}
\usage{
diagnosticar_variable_basica(df, var, nombre_base)
}
\arguments{
\item{df}{Data frame de entrada.}

\item{var}{Nombre de la variable a diagnosticar.}

\item{nombre_base}{Etiqueta de la base revisada.}
}
\value{
Tibble de una fila con metricas basicas de la variable.
}
\description{
Resume clase, tamanio, faltantes, vacios, valores unicos y hasta tres
ejemplos no faltantes de una variable.
}
\examples{
diagnosticar_variable_basica(
  data.frame(x = c("a", "", NA, "b")),
  "x",
  "demo"
)

}
