% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins_capitulos.R
\name{diagnosticar_variables_logicas}
\alias{diagnosticar_variables_logicas}
\title{Diagnosticar variables logicas en capitulos cargados}
\usage{
diagnosticar_variables_logicas(dfs)
}
\arguments{
\item{dfs}{Lista nombrada de data frames de capitulos.}
}
\value{
Tibble con \code{capitulo}, \code{variable} y \code{clase}.
}
\description{
Identifica columnas con clase \code{logical} dentro de una lista de
capitulos. En capitulos crudos cargados desde archivos \code{CAP_*}, una
columna logica suele indicar inferencia automatica de tipos y debe revisarse.
}
\examples{
diagnosticar_variables_logicas(list(A = data.frame(x = c(TRUE, FALSE))))

}
