% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_capitulos.R
\docType{data}
\name{dic_mpios}
\alias{dic_mpios}
\title{Diccionario de municipios de la Encuesta Multipropósito}
\format{
Un tibble con 22 filas y 2 columnas:
\describe{
\item{cod_mpio}{Código del municipio}
\item{municipio}{Nombre del municipio}
}
}
\usage{
dic_mpios
}
\description{
Tabla de correspondencia entre código de municipio y nombre del municipio.
Se usa para enriquecer capítulos con nombres legibles a partir de códigos.
}
\examples{
dic_mpios

}
\keyword{datasets}
