% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.agregar_descripciones_variables_criterios_muestras}
\alias{.agregar_descripciones_variables_criterios_muestras}
\title{Agregar descripciones de diccionario a tabla de variables}
\usage{
.agregar_descripciones_variables_criterios_muestras(tabla)
}
\value{
Tibble con columna \code{descripcion_diccionario}.
}
\description{
Agregar descripciones de diccionario a tabla de variables
}
