% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.cascada_ambito_nivel_muestras}
\alias{.cascada_ambito_nivel_muestras}
\title{Construir cascada para un ambito y nivel}
\usage{
.cascada_ambito_nivel_muestras(
  universo,
  flags,
  municipios,
  ambito,
  municipio,
  municipio_norm,
  nivel,
  tabla_variables_criterios
)
}
\value{
Tibble de siete pasos de cascada.
}
\description{
Construir cascada para un ambito y nivel
}
