% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.cascada_encuestas_muestras}
\alias{.cascada_encuestas_muestras}
\title{Construir cascada ejecutiva de caidas para Muestras}
\usage{
.cascada_encuestas_muestras(
  dfs,
  diag_ref,
  reporte_final,
  municipios,
  tabla_variables_criterios
)
}
\value{
Tibble con cascada por ambito y nivel.
}
\description{
Construir cascada ejecutiva de caidas para Muestras
}
