% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.descripcion_diccionario_variable_muestras}
\alias{.descripcion_diccionario_variable_muestras}
\title{Obtener descripcion de una variable}
\usage{
.descripcion_diccionario_variable_muestras(variable, dic)
}
\value{
Texto de descripcion o fallback.
}
\description{
Obtener descripcion de una variable
}
