% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.detalle_municipios_reporte_muestras}
\alias{.detalle_municipios_reporte_muestras}
\title{Construir detalle municipal para municipios objetivo}
\usage{
.detalle_municipios_reporte_muestras(
  personas_eval,
  reporte_final,
  municipios,
  municipios_objetivo
)
}
\value{
Tibble de detalle municipal.
}
\description{
Helper interno que cruza personas evaluadas con reporte final y municipio.
}
