% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.detalle_variables_componente_cascada_muestras}
\alias{.detalle_variables_componente_cascada_muestras}
\title{Resumir variables de un componente para la cascada}
\usage{
.detalle_variables_componente_cascada_muestras(tabla, componente)
}
\value{
Lista con textos para columnas de cascada.
}
\description{
Resumir variables de un componente para la cascada
}
