% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.diccionario_variables_reporte_muestras}
\alias{.diccionario_variables_reporte_muestras}
\title{Leer diccionario oficial de variables}
\usage{
.diccionario_variables_reporte_muestras()
}
\value{
Tibble con \code{variable} y \code{descripcion}.
}
\description{
Leer diccionario oficial de variables
}
