% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.extraer_reglas_tematica_reporte_muestras}
\alias{.extraer_reglas_tematica_reporte_muestras}
\title{Extraer reglas tematicas desde el diagnostico}
\usage{
.extraer_reglas_tematica_reporte_muestras(
  diag_con_tematica = NULL,
  reglas_tematica = NULL
)
}
\value{
Tibble con \code{capitulo}, \code{nivel}, \code{variable} y \code{descripcion}.
}
\description{
Extraer reglas tematicas desde el diagnostico
}
