% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.extraer_variables_tematica_reporte_muestras}
\alias{.extraer_variables_tematica_reporte_muestras}
\title{Extraer variables tematicas desde el reporte final}
\usage{
.extraer_variables_tematica_reporte_muestras(reporte_final = NULL)
}
\value{
Vector de variables observadas.
}
\description{
Extraer variables tematicas desde el reporte final
}
