% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.filas_tematica_variables_criterios_muestras}
\alias{.filas_tematica_variables_criterios_muestras}
\title{Construir filas de tematica para la tabla de variables}
\usage{
.filas_tematica_variables_criterios_muestras(
  diag_con_tematica = NULL,
  reglas_tematica = NULL,
  reporte_final = NULL
)
}
\value{
Tibble con variables tematicas detectadas.
}
\description{
Construir filas de tematica para la tabla de variables
}
