% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.flags_caida_nivel_muestras}
\alias{.flags_caida_nivel_muestras}
\title{Propagar flags de caida al nivel solicitado}
\usage{
.flags_caida_nivel_muestras(reporte_final, nivel)
}
\value{
Tibble con llaves del nivel y flags logicos.
}
\description{
Propagar flags de caida al nivel solicitado
}
