% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.inferir_capitulo_variable_reporte_muestras}
\alias{.inferir_capitulo_variable_reporte_muestras}
\title{Inferir capitulo desde nombre de variable}
\usage{
.inferir_capitulo_variable_reporte_muestras(variable)
}
\value{
Codigo de capitulo o \code{NA_character_}.
}
\description{
Inferir capitulo desde nombre de variable
}
