% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.inferir_nivel_variable_reporte_muestras}
\alias{.inferir_nivel_variable_reporte_muestras}
\title{Inferir nivel desde nombre de variable}
\usage{
.inferir_nivel_variable_reporte_muestras(variable)
}
\value{
Nivel metodologico de la variable.
}
\description{
Inferir nivel desde nombre de variable
}
