% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.observacion_propagacion_cascada_muestras}
\alias{.observacion_propagacion_cascada_muestras}
\title{Describir propagacion de caidas para cascada}
\usage{
.observacion_propagacion_cascada_muestras(nivel)
}
\value{
Texto metodologico.
}
\description{
Describir propagacion de caidas para cascada
}
