% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporte_criterios_completitud_muestras.R
\name{.separar_variables_reporte_muestras}
\alias{.separar_variables_reporte_muestras}
\title{Separar listas de variables reportadas como texto}
\usage{
.separar_variables_reporte_muestras(x)
}
\value{
Vector de nombres de variable.
}
\description{
Separar listas de variables reportadas como texto
}
