% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{make_value_box}
\alias{make_value_box}
\title{Crear una caja de valor para interfaces Shiny}
\usage{
make_value_box(title, value, subtitle = NULL, meta = NULL)
}
\arguments{
\item{title}{Título de la caja.}

\item{value}{Valor principal que se desea mostrar.}

\item{subtitle}{Texto opcional de apoyo.}

\item{meta}{Meta opcional usada para calcular porcentaje de cumplimiento
y color semaforizado.}
}
\value{
Un objeto HTML tipo \code{shiny.tag}.
}
\description{
Construye un bloque visual en HTML para mostrar un título, un valor principal
y, opcionalmente, un subtítulo. Está pensada para usarse en tableros Shiny.
}
\details{
La función usa internamente \code{div()}, \code{tags$div()} y
\code{comma()} para construir y formatear la salida.
}
\examples{
\dontrun{
make_value_box("Hogares completos", 15234, "Corte a febrero")
}

}
