% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizacion_cruces.R
\name{normalizar_texto_cruce}
\alias{normalizar_texto_cruce}
\title{Normalizar texto para cruces entre fuentes}
\usage{
normalizar_texto_cruce(x)
}
\arguments{
\item{x}{Vector de texto.}
}
\value{
Vector character con texto normalizado.
}
\description{
Limpia campos textuales usados en identificacion, como nombres y apellidos:
fuerza UTF-8 con \code{arreglar_utf8()}, convierte a mayusculas, remueve
tildes y deja solo letras y espacios comprimidos.
}
\examples{
normalizar_texto_cruce(c(" Jose  Gomez ", "Maria-Luisa"))

}
