% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{safe_date}
\alias{safe_date}
\title{Convertir fechas a \code{Date} de forma segura}
\usage{
safe_date(x)
}
\arguments{
\item{x}{Vector de entrada con fechas.}
}
\value{
Un vector de clase \code{Date}.
}
\description{
Convierte un vector a clase \code{Date} intentando primero el formato
\code{"\%d/\%m/\%y"} y luego, para los casos no convertidos, el formato
\code{"\%d/\%m/\%Y"}.
}
\details{
Si el objeto ya es de clase \code{Date}, \code{POSIXct} o \code{POSIXt},
simplemente se devuelve como \code{Date}.
}
\examples{
safe_date(c("01/02/25", "15/03/2025"))

}
